+++ Assets/Ipocom/Runtime/UdpClientStream.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1dc0f0f925cfb942a50c7fbf46168cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
