+++ Assets/Ipocom/Runtime/UdpSocketStream.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Ipocom
+{
+    public class UdpSocketStream : IDisposable
+    {
+        public void Dispose()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
