+++ Assets/Ipocom/Runtime/UdpSocketStream.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc01d16f7ae1c7f49b3bf5bf37c5913f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
