+++ Packages/manifest.json
@@ -1,8 +1,8 @@
 {
   "dependencies": {
-    "com.unity.collab-proxy": "1.15.18",
-    "com.unity.ide.rider": "2.0.7",
-    "com.unity.ide.visualstudio": "2.0.15",
+    "com.unity.collab-proxy": "1.17.2",
+    "com.unity.ide.rider": "3.0.15",
+    "com.unity.ide.visualstudio": "2.0.16",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.test-framework": "1.1.31",
     "com.unity.textmeshpro": "3.0.6",
